remove unused release functionality
